+++ buildSrc/build.gradle.kts
@@ -11,7 +11,6 @@ repositories {
 dependencies {
     implementation(kotlin("gradle-plugin", version = "1.3.72"))
     implementation("com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.5")
-    implementation(kotlin("gradle-plugin", version = "1.3.71"))
     implementation("com.fasterxml.jackson.core:jackson-databind:2.11.0")
 
     //Remember to change version according to DependenciesVersion
